Updates electionId parameter type from Long to Int in getVoterInfo()

+++ app/src/main/java/com/example/android/politicalpreparedness/network/CivicsApiService.kt
@@ -43,7 +43,7 @@ interface CivicsApiService {
     @GET("voterinfo")
     suspend fun getVoterInfo(
             @Query("address") address: String,
-            @Query("electionid") electionId: Long
+            @Query("electionid") electionId: Int
     ) : VoterInfoResponse
 
     //COMPLETE: Add representatives API Call
